events WIP; stashish for merging NADGITs PR
